[telepathy-gabble] Fix running some tests. Contributes to JB#42708
All of these were failing on readlink parameter busybox doesn't
support. Changed that.

For rest the python3 port seems halfway there, failing a lot on
bytes vs strings typing. Fixed some of those, not sure would it be
worth the effort in handling all if we don't really do telepathy-gabble
development.

Now 132 out of 245 pass which is already quite a lot better.